# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "c-plane API",
    "description": "Control plane API for managing projects, environments, containers, and databases",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/organization/{organization_id}/databases/postgres": {
      "get": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "list_databases",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project databases with branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatabaseWithBranchesResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:read"
      },
      "post": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "create_database",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Database created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:create"
      }
    },
    "/api/organization/{organization_id}/databases/postgres/{database_id}": {
      "get": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "get_database",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database_id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:read"
      },
      "delete": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "delete_database",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database_id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database deleted"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:delete"
      },
      "patch": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "update_database",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database_id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Database updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:update"
      }
    },
    "/api/organization/{organization_id}/databases/postgres/{database_id}/branches": {
      "get": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "list_database_branches",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database_id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database branch links",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatabaseBranchResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:read"
      },
      "post": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "create_database_branch",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database_id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Link already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseBranchResponse"
                }
              }
            }
          },
          "201": {
            "description": "Database branch link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseBranchResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:manage"
      }
    },
    "/api/organization/{organization_id}/databases/postgres/{database_id}/branches/{branch_id}": {
      "delete": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "delete_database_branch",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database_id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "branch_id",
            "in": "path",
            "description": "Branch ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database branch link deleted"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:delete"
      },
      "patch": {
        "tags": [
          "databases/postgres"
        ],
        "operationId": "update_database_branch",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "database_id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "branch_id",
            "in": "path",
            "description": "Branch ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatabaseBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Database branch updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseBranchResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "database:postgres:manage"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDatabaseBranchRequest": {
        "type": "object",
        "required": [
          "branch_id"
        ],
        "properties": {
          "autoscaling_enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "autoscaling_max_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "autoscaling_min_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "backup_retention_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "high_availability": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "ram": {
            "type": [
              "string",
              "null"
            ]
          },
          "read_replicas": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "CreateDatabaseRequest": {
        "type": "object",
        "required": [
          "name",
          "project_id",
          "region_id"
        ],
        "properties": {
          "autoscaling_enabled": {
            "type": "boolean"
          },
          "autoscaling_max_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "autoscaling_min_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "backup_retention_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "high_availability": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "ram": {
            "type": [
              "string",
              "null"
            ]
          },
          "read_replicas": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "region_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "DatabaseBranchResponse": {
        "type": "object",
        "required": [
          "id",
          "database_id",
          "branch_id",
          "organization_id",
          "high_availability",
          "autoscaling_enabled",
          "organization_region_backup_bucket_id",
          "backup_credential_id",
          "backup_prefix"
        ],
        "properties": {
          "autoscaling_enabled": {
            "type": "boolean"
          },
          "autoscaling_max_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "autoscaling_min_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "backup_credential_id": {
            "type": "string",
            "format": "uuid"
          },
          "backup_prefix": {
            "type": "string"
          },
          "backup_retention_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "database_id": {
            "type": "string",
            "format": "uuid"
          },
          "high_availability": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_region_backup_bucket_id": {
            "type": "string",
            "format": "uuid"
          },
          "ram": {
            "type": [
              "string",
              "null"
            ]
          },
          "read_replicas": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "DatabaseResponse": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "region_id"
        ],
        "properties": {
          "default_branch_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "region_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "DatabaseWithBranchesResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DatabaseResponse"
          },
          {
            "type": "object",
            "required": [
              "branches"
            ],
            "properties": {
              "branches": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DatabaseBranchResponse"
                }
              }
            }
          }
        ]
      },
...
      "UpdateDatabaseBranchRequest": {
        "type": "object",
        "properties": {
          "autoscaling_enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "autoscaling_max_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "autoscaling_min_cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "backup_retention_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "cpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "high_availability": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "ram": {
            "type": [
              "string",
              "null"
            ]
          },
          "read_replicas": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "UpdateDatabaseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key with the operation's required scope."
      },
      "registryBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "serviceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cplane-token",
        "description": "Internal service token."
      }
    }
  },
...
}